Fix exception handling in the very initial stages of ORCA optimization
When ORCA runs out of memory while setting up its memory pool manager, in
CMemoryPoolManager::SetupGlobalMemoryPoolManager<>, it used to ignore
exceptions other than those of minor type ExmiOOM. When we use the
default GPDB memory allocators (optimizer_use_gpdb_allocators = true),
then an OOM condition will generate a different type of exception
(minor type of ExmiGPDBError). Ignoring that exception leads to a
crash with a segment violation.

The fix is to re-throw either of these OOM exceptions, so that they
are caught in the try/catch block in global method InitGPOPT(). Since
this is called before we set up a longjump buffer, that still causes a
FATAL error and an abort, but that's an improvement over the crash.

The reason for changing the behavior for exception of type ExmiOOM is
that the calling methods currently continue for any type of exception
seen in CMemoryPoolManager::SetupGlobalMemoryPoolManager<>, which
will cause a crash.