Improve handling of failures in the InitDXL method
This method has a variable that prevents initializing the DXL context
more than once. Set this variable only upon exit from the method, so
that we re-attempt the code on the next query when we fail (at least
some failures will cause a longjmp to a higher stack frame and won't
take the regular exit path). Note that there is no guarantee that it
will work the second time, expecially when the failure was inside the
xerces initialization code (detected or undetected).  For simple
memory allocation failures in GPDB code it should work, though. Note
also that failures could cause small memory leaks, which is probably
acceptable and common in most of our code.